Easier testing with aula client in seperate file (#5)
* Moved aula client to own file
