+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "node-red-contrib-aula-education",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "A Node-Red Node for collecting data from aula the danish educational platform.",
   "keywords": [
     "node-red",
@@ -15,15 +15,20 @@
     "name": "zinen"
   },
   "files": [
+    "aula.js",
     "aula.html",
     "examples/"
   ],
-  "main": "aula.js",
+  "main": "aula-client.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/zinen/node-red-contrib-aula-education"
   },
-  "scripts": {"test": "standard"},
+  "scripts": {
+    "test": "standard",
+    "lint": "standard --fix",
+    "dev-start": "node ./test/test.dev.js"
+  },
   "node-red": {
     "version": ">=1.0.0",
     "nodes": {
